refactor: move Icone para util
